# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API para radioxenter",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module."
  },
  "host": "localhost:4000",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/convenios/{cod_convenio}": {
      "get": {
        "tags": [
          "Convenios"
        ],
        "description": "Endpoint para obtener un convenio",
        "parameters": [
          {
            "name": "cod_convenio",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/convenios": {
      "get": {
        "tags": [
          "Convenios"
        ],
        "description": "Endpoint para obtener convenios",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Convenios"
        ],
        "description": "Endpoint para registrar un convenio.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Convenio"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "Convenios"
        ],
        "description": "Endpoint para editar un convenio.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Convenio"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/entidades/{cod_entidad}": {
      "get": {
        "tags": [
          "Entidades"
        ],
        "description": "Endpoint para obtener una entidad",
        "parameters": [
          {
            "name": "cod_entidad",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/entidades": {
      "get": {
        "tags": [
          "Entidades"
        ],
        "description": "Endpoint para obtener entidades",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Entidades"
        ],
        "description": "Endpoint para crear una entidad.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Entidad"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "Entidades"
        ],
        "description": "Endpoint para editar una entidad.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Entidad"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/entidades/{cod_entidad}/convenios": {
      "get": {
        "tags": [
          "Entidades"
        ],
        "description": "Endpoint para obtener todos los convenios de una entidad",
        "parameters": [
          {
            "name": "cod_entidad",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Entidades"
        ],
        "description": "Endpoint para editar los convenios de una entidad.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EditarConvenioEntidad"
            }
          },
          {
            "name": "cod_entidad",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/entidades/{cod_entidad}/convenios/{cod_servicio}": {
      "put": {
        "tags": [
          "Entidades"
        ],
        "description": "Endpoint para editar un servicio de un convenio con una entidad.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ConvenioUpdate"
            }
          },
          {
            "name": "cod_entidad",
            "in": "path",
            "required": true
          },
          {
            "name": "cod_servicio",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Entidad": {
      "type": "object",
      "properties": {
        "cod_entidad": {
          "type": "number",
          "example": 0
        },
        "razon_social_entidad": {
          "type": "string",
          "example": ""
        },
        "nombre_comercial_entidad": {
          "type": "string",
          "example": ""
        },
        "nit_entidad": {
          "type": "string",
          "example": ""
        },
        "direccion_entidad": {
          "type": "string",
          "example": ""
        },
        "telefono_entidad": {
          "type": "string",
          "example": ""
        },
        "nombre_representante": {
          "type": "string",
          "example": ""
        },
        "cedula_representante": {
          "type": "number",
          "example": 0
        },
        "telefono_representante": {
          "type": "string",
          "example": ""
        },
        "correo_representante": {
          "type": "string",
          "example": "email@email.com"
        },
        "nombre_contacto": {
          "type": "string",
          "example": ""
        },
        "cedula_contacto": {
          "type": "string",
          "example": ""
        },
        "telefono_contacto": {
          "type": "string",
          "example": ""
        },
        "correo_contacto": {
          "type": "string",
          "example": ""
        },
        "cod_forma_de_pago_entidad": {
          "type": "number",
          "example": 0
        },
        "cod_tipo_facturacion": {
          "type": "number",
          "example": 0
        }
      },
      "xml": {
        "name": "Entidad"
      }
    },
    "EditarConvenioEntidad": {
      "type": "object",
      "properties": {
        "servicios_convenio": {
          "type": "array",
          "items": {
            "type": "number"
          }
        },
        "valores_servicios": {
          "type": "array",
          "items": {
            "type": "number"
          }
        },
        "fecha_inicial_convenio": {
          "type": "string",
          "example": "dd/MM/YYYYY"
        },
        "fecha_final_convenio": {
          "type": "string",
          "example": "dd/MM/YYYYY"
        }
      },
      "xml": {
        "name": "EditarConvenioEntidad"
      }
    },
...
    "Convenio": {
      "type": "object",
      "properties": {
        "valor_servicio": {
          "type": "number",
          "example": 0
        },
        "fecha_inicial_convenio": {
          "type": "string",
          "example": "yyyy-MM-dd"
        },
        "fecha_final_convenio": {
          "type": "string",
          "example": "yyyy-MM-dd"
        },
        "cod_entidad": {
          "type": "number",
          "example": 0
        },
        "cod_servicio": {
          "type": "number",
          "example": 0
        }
      },
      "xml": {
        "name": "Convenio"
      }
    },
    "ConvenioUpdate": {
      "type": "object",
      "properties": {
        "valor_servicio": {
          "type": "number",
          "example": 0
        },
        "fecha_inicial_convenio": {
          "type": "string",
          "example": "yyyy-MM-dd"
        },
        "fecha_final_convenio": {
          "type": "string",
          "example": "yyyy-MM-dd"
        }
      },
      "xml": {
        "name": "ConvenioUpdate"
      }
    },
...
  }
}